+++ examples/shared/xy-drawing-series-with-mouse-or-touch/index.ts
@@ -2,15 +2,18 @@ import * as am5 from "@amcharts/amcharts5";
 import * as am5xy from "@amcharts/amcharts5/xy";
 import am5themes_Animated from "@amcharts/amcharts5/themes/Animated";
 
+
 // Create root element
 // https://www.amcharts.com/docs/v5/getting-started/#Root_element
 let root = am5.Root.new("chartdiv");
+
 
 // Set themes
 // https://www.amcharts.com/docs/v5/concepts/themes/
 root.setThemes([
   am5themes_Animated.new(root)
 ]);
+
 
 // Create chart
 // https://www.amcharts.com/docs/v5/charts/xy-chart/
@@ -25,9 +28,14 @@ let chart = root.container.children.push(
 
 chart.get("colors").set("step", 3);
 
+
+// Add cursor
+// https://www.amcharts.com/docs/v5/charts/xy-chart/cursor/
 let cursor = chart.set("cursor", am5xy.XYCursor.new(root, {}));
 cursor.lineY.set("visible", false);
 
+
+// Generate random data
 let date = new Date();
 date.setHours(0, 0, 0, 0);
 let value = 100;
@@ -35,7 +43,10 @@ let value = 100;
 function generateData() {
   value = Math.round((Math.random() * 10 - 5) + value);
   am5.time.add(date, "day", 1);
-  return { date: date.getTime(), value: value };
+  return {
+    date: date.getTime(),
+    value: value
+  };
 }
 
 function generateDatas(count) {
@@ -46,14 +57,21 @@ function generateDatas(count) {
   return data;
 }
 
+
 // Create axes
 // https://www.amcharts.com/docs/v5/charts/xy-chart/axes/
 let xAxis = chart.xAxes.push(
   am5xy.DateAxis.new(root, {
     maxDeviation: 0.3,
-    baseInterval: { timeUnit: "day", count: 1 },
+    baseInterval: {
+      timeUnit: "day",
+      count: 1
+    },
     renderer: am5xy.AxisRendererX.new(root, {}),
-    tooltip: am5.Tooltip.new(root, { themeTags: ["axis"], animationDuration: 200 })
+    tooltip: am5.Tooltip.new(root, {
+      themeTags: ["axis"],
+      animationDuration: 200
+    })
   })
 );
 
@@ -65,26 +83,55 @@ let yAxis = chart.yAxes.push(
 );
 
 
-let series = chart.series.push(am5xy.LineSeries.new(root, { name: "Series", xAxis: xAxis, yAxis: yAxis, valueYField: "value", valueXField: "date" }));
+let series = chart.series.push(am5xy.LineSeries.new(root, {
+  name: "Series",
+  xAxis: xAxis,
+  yAxis: yAxis,
+  valueYField: "value",
+  valueXField: "date"
+}));
 
-let drawingSeries = chart.series.push(am5xy.LineSeries.new(root, { name: "Series", xAxis: xAxis, yAxis: yAxis, valueYField: "value", valueXField: "date" }));
+let tooltip = series.set("tooltip", am5.Tooltip.new(root, {}));
+tooltip.label.set("text", "{valueY}");
 
-// invisible bullet which will be dragged (to avoid some conflicting between drag position and bullet position which results flicker)
-drawingSeries.bullets.push(function () {
-  let bulletCircle = am5.Circle.new(root, { radius: 6, fillOpacity:0, fill: drawingSeries.get("fill"), draggable: true, cursorOverStyle:"pointer" });
-  bulletCircle.events.on("dragged", function (e) {
+let drawingSeries = chart.series.push(am5xy.LineSeries.new(root, {
+  name: "Series",
+  xAxis: xAxis,
+  yAxis: yAxis,
+  valueYField: "value",
+  valueXField: "date"
+}));
+
+// Invisible bullet which will be dragged (to avoid some conflicting between
+// drag position and bullet position which results flicker)
+drawingSeries.bullets.push(function() {
+  let bulletCircle = am5.Circle.new(root, {
+    radius: 6,
+    fillOpacity: 0,
+    fill: drawingSeries.get("fill"),
+    draggable: true,
+    cursorOverStyle: "pointer"
+  });
+  bulletCircle.events.on("dragged", function(e) {
     handleDrag(e);
   })
-  return am5.Bullet.new(root, { sprite: bulletCircle })
+  return am5.Bullet.new(root, {
+    sprite: bulletCircle
+  })
 })
 
-// actual bullet
-drawingSeries.bullets.push(function () {
-  let bulletCircle = am5.Circle.new(root, { radius: 5, fill: drawingSeries.get("fill") });
-  return am5.Bullet.new(root, { sprite: bulletCircle })
+// Actual bullet
+drawingSeries.bullets.push(function() {
+  let bulletCircle = am5.Circle.new(root, {
+    radius: 5,
+    fill: drawingSeries.get("fill")
+  });
+  return am5.Bullet.new(root, {
+    sprite: bulletCircle
+  })
 })
 
-// drag handler
+// Drag handler
 function handleDrag(e) {
   let point = chart.plotContainer.toLocal(e.point);
   let date = xAxis.positionToValue(xAxis.coordinateToPosition(point.x));
@@ -97,8 +144,6 @@ function handleDrag(e) {
   dataItem.set("valueYWorking", value);
 }
 
-let tooltip = series.set("tooltip", am5.Tooltip.new(root, {}));
-tooltip.label.set("text", "{valueY}");
 
 // Add scrollbar
 // https://www.amcharts.com/docs/v5/charts/xy-chart/scrollbars/
@@ -106,26 +151,33 @@ chart.set("scrollbarX", am5.Scrollbar.new(root, {
   orientation: "horizontal"
 }));
 
+
+// Set data
 let data = generateDatas(1200);
 series.data.setAll(data);
+
 
 // Make stuff animate on load
 // https://www.amcharts.com/docs/v5/concepts/animations/
 series.appear(1000);
 chart.appear(1000, 100);
 
-// interactivity
-chart.plotContainer.get("background").events.on("click", function (e) {
+// Interactivity
+chart.plotContainer.get("background").events.on("click", function(e) {
   let point = chart.plotContainer.toLocal(e.point);
   let date = xAxis.positionToValue(xAxis.coordinateToPosition(point.x));
   let value = yAxis.positionToValue(yAxis.coordinateToPosition(point.y));
-  drawingSeries.data.push({ date: date, value: value });
+  drawingSeries.data.push({
+    date: date,
+    value: value
+  });
 
   sortData();
 })
 
 
-// sort data so that if clicked between existing data items, the item would be added between
+// Sort data so that if clicked between existing data items, the item would
+// be added between
 function sortData() {
   drawingSeries.dataItems.sort(function(a, b) {
     let atime = a.get("valueX");
@@ -133,15 +185,17 @@ function sortData() {
 
     if (atime < btime) {
       return -1;
-    }
-    else if (atime == btime) {
+    } else if (atime == btime) {
       return 0;
-    }
-    else {
+    } else {
       return 1;
     }
   })
 }
 
-// explanatory labels
-chart.plotContainer.children.push(am5.Label.new(root, {x:10, y:10,  text:"Click on plot area to draw a series"}));
+// Explanatory labels
+chart.plotContainer.children.push(am5.Label.new(root, {
+  x: 10,
+  y: 10,
+  text: "Click on plot area to draw a series"
+}));